fix: workaround for duplicate keys when adding new documents to Open Zaak  (#171)
Workaround for duplicate keys when adding new documents to Open Zaak by
using '999' ids for inserted document rows in the Open Zaak database. It
appears that Open Zaak uses its cache for this data and does not see
this data that was inserted directly in the database at startup.

Solves PZ-700.