feat: grant the beheerder rol access to the 'werklijst zaken-taken' (#595)
Grant the beheerder rol access to the 'werklijst zaken-taken' permission
so that this role can use the ZAC dashboard. This as a workaround for
the issue where users with only the beheerder role could not use the
dashboard nor ZAC itself (500 error) because they were missing this
permission. #minor

Solves PZ-1499