chore: generate Java model classes for new 'productaanvraag dimpact' object type (#701)
Added OpenAPI spec for new 'productaanvraag dimpact' object type so that
Java model classes for this object type are generated. Added unit test
for the NotifcatieReceiver class. And also moved these generated Java
classes to a 'generated' subpackage. Note that in this PR this new
productaanvraag type is not actually used yet. It is just preparing for
that.

Solves PZ-1477