fix(itest): ignore response data in the test for removal of of a zaaklocatie
OpenZaak is not processing the removal of the zaaklocatie as expected in the integration test setup. We will therefore not check the updated zaak when we remove a zaaklocatie.

The wildfly configuration script has been updated so that we can easily enable request logging.

Fixed a reference to a non-existing buildZacDockerImage gradle goal in the testing documentation.