feat: remove kvk search on rsin (#1621)
Remove KVK search on RSIN and removed related RSIN and rechtspersoon
code, both from frontend and backend. This because rechtspersonen were
not correctly implemented and incorrectly treated as a kind of company,
but they are something else and this resulted in errors when a single
RSIN has multiple KVK numbers.

Solves PZ-3631