feat: FE align body styles to Figma design pattern settings (#1976)
FE: align body styles to current Figma design pattern settings; and
apply same styles to the SD wizard result page. The font-display swap is
adviced for a fast and smooth transition after the Roboto is loaded.
Will most probably be just the very first time the Behandelaar visits
the app, since cached by the browser (if not already).

Solves: PZ-4348