feat: FE zaken signalering - paginated requests (#2017)
With this PR the web socket providing all the open signals is replaced
by paginated requests. On navigating to a next page a new request to the
server is executed. And when a general signalering signal arrives, the
current page is refetched. Making a new signal shift the whole list.
This shifting behaviour is identical to the previous web socket solution
behaviour.

Solves: PZ-2612