feat: FE - SmartDocuments form functions tests (#2155)
FE - SmartDocuments form tests

- Added tests for all array functions
- Redone function flattenNestedGroups since old one flattend but as well
left nested branches in its result
- Removed redundant flatten function for its newer and leaner brother
- Stripped the flattened groups from group(s) that don't have any
templates (array length === 0). In case of editing SmartDocuments form
of no influence since ignored anyway, BUT (!!) in case of Document
create (from Zaak or Taak) now not showing a Group anymore that has no
Templates to show. This for sure will be a future finding as it is now
one of my findings!

Solves PZ-4618