feat(pages): build github pages documentation
This is a simple setup for generating GitHub pages for the project. It uses the GitHub Jekyll setup for generating a static site. In our case there are sources for the basic static site in src/main/pages, and a script that will pull some of the build generated content in script/pages.

Changes made to the build process are:
- Gradle build will generate HTML content when running in CI
- Actions workflow build-test-deploy now has extra jobs to generate the pages