fix: zaak geometry should not be deleted on zaak update (#2588)
Fixed issue where zaak geometry was deleted on every zaak update and
fixed issue where zaak geometry update and deletions sometimes worked
and sometimes did not. The issue was that for deletions a @JsonbNillable
annotation was used on the Zaak object in the ZrcClient but for normal
updates not. However this annotation is static and there is only one
zaak patch method in the ZrcClient and this resulted in inconsistent
behaviour. This was solved by introducing a custom JSONB serializer for
the zaak Geometry class that dynamically either writes a null (for
deletions) or the geometry (Point only for now) object to JSON.

Solves  PZ-5299

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Hristo Iliev <hsiliev@gmail.com>