fix: undo moving of classes used in our CMMN model (#2675)
Undo moving of classes used in our CMMN model because this breaks all
zaken and tasks that were created by earlier versions of the CMMN model!
Do note that this change will again break those zaken and tasks that
were created by the previous version of the CMMN model (luckily that
only applies to the INFO test environment).

Solves PZ-5466