fix: e2e test (#3206)
This pull request updates a test step definition in the
`src/e2e/step-definitions/external-systems.ts` file to improve the
reliability and clarity of element selection in the test.

Improvements to element selection:

* Replaced the use of `getByText` with `getByRole` for selecting the
"Create Document" button. This change makes the selection more robust
and accessible by targeting the button role with a regex for its name.
* Updated the logic for waiting for the "Document maken" sidebar to
appear by using `getByRole` to target a heading with a specific name.
This improves clarity and ensures the test waits for the correct
element. And is according to the [testing librarty
priority](https://testing-library.com/docs/queries/about/#priority)

Solves broken E2E test