fix(app): Opgeschorte Zaak - Aanvullende informatie form - Show hervatten options (#3212)
FE - Opgeschorte Zaak - Aanvullende informatie form - Show hervatten
options

Strange thing is that the show/hide of this options field depended on a
certain indication in the Zaak indications. When a Zaak is opgeaschort
there is the 'opgeschort' indication in the array. But on a Zaak which
is 'gekoppeld' there is only the 'DEELZAAK' or 'HOOFDZAAK' indication in
the array. No sign of 'opgeschort' anymore. That might be a bug in
itself.

On the other hand, TMHO we should not depend on an array when there is
the prop 'isOpgeschort' on the Zaak object itzelf. So this bug can be
solved easily. The possible indications array error needs to be
investigated.

Solves PZ-6546