fix(zaken.service): call correct endpoint (#3332)
This pull request includes a small but important fix in the
`ZakenService` class to correct the API endpoint used for releasing
cases from a list.

*
[`src/main/app/src/app/zaken/zaken.service.ts`](diffhunk://#diff-77edb5ea0afd29dea13d23b27f4e2d2a3c854f5e995d92ce1f780cfb580ee286L196-L199):
Updated the endpoint in the `vrijgevenVanuitLijst` method from
`"/rest/zaken/lijst/verdelen"` to `"/rest/zaken/lijst/vrijgeven"` to
align with the correct API functionality.

Solves PZ-6713