test(E2E): validate that the "no acknowledgement" indication is set (#3337)
This pull request introduces a new feature to the end-to-end (E2E) tests
for verifying that a user can see an indication when no acknowledgment
has been sent for a "zaak" (case). The changes include updates to the
feature file and the corresponding step definitions.

### Updates to E2E tests:

*
[`src/e2e/features/2-zaken-toevoegen.feature`](diffhunk://#diff-9179cb8d57f529f7fb196d94e12294b29958874368cbb8b0e11c5ac9ea14d5b0R12):
Added a new scenario step to check that "Bob" sees an indication when no
acknowledgment has been sent for a created zaak.

*
[`src/e2e/step-definitions/zaak.ts`](diffhunk://#diff-14162a669617ec1a3d444c5fbd477deb3e031abda24ac142b631cf818cf9ccffR259-R268):
Implemented a new step definition to verify the visibility of the "Geen
bevestiging verstuurd" (No acknowledgment sent) indication using the
`getByRole` method.

Solves PZ-5641