feature(action): action to create a minor release (#3451)
Updated the manual trigger (workflow_dispatch) for the build-test-deploy
github action workflow.

The manual trigger now requires you to choose the type of version bump
you need for this build, options are: major, minor, patch, defaulting to
patch.

The workflow will then verify that, the manual trigger is only ever run
on the `main` branch. If it is not, then the build fails.
If it is, then the build will use the selected version bump for
determining the next application version.

If the workflow is not triggered manually, the version bump defaults to
patch per normal behaviour.

Solves: PZ-6591