chore(zaak-view): put all `indicaties` on a separate row (#3499)
This pull request modifies the layout and structure of the
`zaak-view.component.html` file to improve the visual organization of
the `mat-card-header` section. The changes primarily focus on
restructuring the header content and adjusting its styling.

### Layout improvements:

*
[`src/main/app/src/app/zaken/zaak-view/zaak-view.component.html`](diffhunk://#diff-77b39b0f489ba66cfa831cae58da307ced947da86c4ff6982a2fd212c09db51cL181-R205):
Changed the `mat-card-header` layout from a `flex-row space-between`
structure to a `flex-col gap-8 space-between` structure. This improves
spacing and alignment for the case `indicaties`.

Solves PZ-6806