chore(DialogData): pass action context to dialog (#3584)
This pull request introduces changes to improve the handling of "unlink"
actions for various entities in the application. Key updates include the
addition of `confirmButtonActionKey` properties for dialog
configurations and corresponding translations for these actions in both
English and Dutch.

### Dialog Configuration Updates:
* Added `confirmButtonActionKey: "actie.document.ontkoppelen"` to the
dialog configuration in `ZaakDocumentenComponent` for unlinking
documents.
* Added `confirmButtonActionKey: "actie.initiator.ontkoppelen"` to the
dialog configuration in `ZaakViewComponent` for unlinking initiators.
* Added `confirmButtonActionKey: "actie.betrokkene.ontkoppelen"` to the
dialog configuration in `ZaakViewComponent` for unlinking stakeholders.

### Translation Updates:
* Added the English translation `"actie.initiator.ontkoppelen": "Unlink
initiator"` to `en.json`.
* Added the Dutch translation `"actie.initiator.ontkoppelen": "Initiator
ontkoppelen"` to `nl.json`.

Solves PZ-6936