fix(frontend): Pass hard-coded action and context to PersoonZoekComponent from zoek.component.html
The PersoonZoekComponent requires both action and context to be passed in. The zoek.component.html therefore has to pass this information. This zoek component is used for a generic search, outside of the context of a task or case, and can therefore set both static context and action.

Solves: PZ-7262