feat: add test case with exception flow (#3563)
Testing of the BRP Protocollering flow udates

A test case has been created to verify the working of the backend code
for when an exception is thrown on fetching a case. When that happens
the business logic will continue and fetch BRP information using the
default configured context.

In the docker-compose setup we introduced a wiremock proxy that
validates the BRP Protocollering headers (minimally for now). When no or
incorrect header values are passed it will fail the BRP call. With
correct headers it will pass the call on to the BRP Personen Mock
service.

During testing a small fix was made to the zoek.component.html; it has
to pass both action and context to PersoonZoekComponent. This zoek
component is used for a generic search, outside of the context of a task
or case, and can therefore set both static context and action.

Solves PZ-7188

---------

Co-authored-by: Hristo Iliev <hristo.iliev@info.nl>
Co-authored-by: Bas de Wit <bas@info.nl>