Fix: return proper error message when attempting to abort a zaak that already has a decision set. (#3714)
<img width="1020" alt="image"
src="https://github.com/user-attachments/assets/05f5c311-8ee4-404a-897f-368b30199c1c"
/>
As per the story comment, a zaak with a decision (resultaat) should not
be allowed to be aborted.
This PR enforces that rule by returning a clear error message when an
abort is attempted on such a zaak.
Solves PZ-5446