refactor(signaleringen-settings-beheer.service): use `GeneratedType`s (#3759)
This pull request refactors the codebase to replace deprecated patterns
and improve maintainability. The most important changes include
replacing the `HttpClient` with `ZacHttpClient` in the
`SignaleringenSettingsBeheerService` and removing the deprecated
`DatumVeld` enum.

### Refactoring to use `ZacHttpClient`:

*
[`src/main/app/src/app/admin/signaleringen-settings-beheer.service.ts`](diffhunk://#diff-8ea1b37ce245444923f766c5a2013a9bcacdec07230eb8d621f68cec4f18f091L6-R29):
Replaced `HttpClient` and `FoutAfhandelingService` with `ZacHttpClient`
for HTTP requests, simplifying error handling and request structure.
Updated the `list` and `put` methods to use `ZacHttpClient.GET` and
`ZacHttpClient.PUT` respectively.

### Removal of deprecated code:

*
[`src/main/app/src/app/zoeken/model/datum-veld.ts`](diffhunk://#diff-eeca6fd9fff995e997e743c936decc2d023e903bdb2b0b42029465f736d05c1bL1-L29):
Removed the deprecated `DatumVeld` enum, as it was not used in the
codebase.