chore: convert mailtemplates package to kotlin (#3832)
Converted (most of the) mailtemplates package to Kotlin. In the end
decided not to try to convert the two remaining model classes because I
was struggling trying to convert them. Maybe for another day.. ;-) There
is still a lot to improve in this code but I will leave that for another
day.. First we should add more test coverage here I think.

Solves PZ-7619