added some more safeguards (#3898)
This pull request introduces improved error handling for session
management and updates the `IndexingService` to handle specific
exceptions more effectively. Key changes include the addition of a
custom exception, `OidcSessionException`, and modifications to ensure
better handling of invalid or expired sessions.

### Error handling improvements:

*
[`src/main/kotlin/nl/info/zac/authentication/OidcSessionService.kt`](diffhunk://#diff-0c5b1e18a964062c750f41f5340a542cf06c8eecc7ca58009673bf8893e82f96L39-R43):
Updated the `refreshUserSession` method to catch `IllegalStateException`
when retrieving the refresh token from the session. If the session is
invalid or expired, an `OidcSessionException` is thrown with a
descriptive error message.

*
[`src/main/kotlin/nl/info/zac/search/IndexingService.kt`](diffhunk://#diff-9b0811e3e3103c5f43c50ad65b5b3a99d623477492e9f5069229f2c9c98e7fa8R375-R381):
Enhanced the `IndexingService` to handle `OidcSessionException` during
session revalidation. If revalidation fails, a warning is logged,
providing better insight into the failure.

### Dependency updates:

*
[`src/main/kotlin/nl/info/zac/search/IndexingService.kt`](diffhunk://#diff-9b0811e3e3103c5f43c50ad65b5b3a99d623477492e9f5069229f2c9c98e7fa8R25):
Added imports for `OidcSessionException`.
[[1]](diffhunk://#diff-9b0811e3e3103c5f43c50ad65b5b3a99d623477492e9f5069229f2c9c98e7fa8R25)

Solves [PZ-6628](https://dimpact.atlassian.net/browse/PZ-6628)