chore(heading-form-field): fix lint issues (#3946)
This pull request introduces default values for properties in the
`HeadingFormField` and `HeadingComponent` classes to improve
initialization consistency and reduce potential runtime errors.

### Default value additions:

*
[`src/main/app/src/app/shared/material-form-builder/form-components/heading/heading-form-field.ts`](diffhunk://#diff-470e8f5f14829b1c39fe99e7db4538844a6e156fbec88d70494ae7f55b12d234L12-R12):
Set a default value for the `level` property in the `HeadingFormField`
class to `HeadingLevel.H1`.
*
[`src/main/app/src/app/shared/material-form-builder/form-components/heading/heading.component.ts`](diffhunk://#diff-55be6deabd012fff98ce361932be7acacf2a4fc9b054aa6c60082d0196d1ab3dL17-R17):
Initialized the `data` property in the `HeadingComponent` class with a
new instance of `HeadingFormField`.

Solves PZ-7675