chore: when handling a productaanvraag start the CMMN process directly after creating the zaak (#3983)
When handling a productaanvraag, start the CMMN process directly after
creating the zaak and only then perform additional logic such as adding
betrokkenen and attachments, so that should the latter fail at least the
CMMN process has been started and the zaak can be handled. Note that
error handling still needs improvement here so this is just a first
small step.

Solves PZ-7885