feat(KlantRestService): pass KvK number when request searches for it (#4028)
This pull request includes a modification to the `KlantRestService`
class to enhance the mapping logic for `RestBedrijf` objects by
including the `kvkNummer` parameter in the transformation process.

### Key change:
*
[`src/main/kotlin/nl/info/zac/app/klant/KlantRestService.kt`](diffhunk://#diff-0e0ebf0b435a67dc6e43acf2c662ecce4bf30e82b65470d9e554285f2655f9bdL177-R177):
Updated the `map` function in the `listBedrijven` method to call
`toRestBedrijf` with an additional `kvkNummer` parameter from
`restListBedrijvenParameters`, enabling more context-aware
transformation.

Solves PZ-6997