fix(besluit-view): use correct translation string (#4059)
This pull request makes a minor update to the confirmation dialog
messaging in the `BesluitViewComponent`. Specifically, it changes the
translation key used for the withdrawal message to simplify or correct
the displayed text.

* Updated the `melding` property in the confirmation dialog
configuration to use the `msg.besluit.intrekken` translation key instead
of `msg.besluit.intrekken.melding` in `besluit-view.component.ts`.