fix: fixed linting errors in websocket folder (#4176)
This PR focusses on solving the typescript linting errors using [this
page](https://dimpact.atlassian.net/wiki/spaces/PZW/pages/357171549/Typescript+issues+ZAC+frontend)
as reference. It focusses solely on the websocket folder of the frontend
and fixes all linting error within that folder.

The following changes have been made
-
[src/main/app/src/app/core/websocket/model/event-suspension.ts](https://github.com/infonl/dimpact-zaakafhandelcomponent/compare/fix/PZ-8160-lint-errors?expand=1#diff-f2cba347a119dc85568bdf0e702c61d9d09423f68970d6833c5721471c71f569):
Added a checkmark to expires since the value will always be set
-
[src/main/app/src/app/core/websocket/model/screen-event-id.ts](https://github.com/infonl/dimpact-zaakafhandelcomponent/compare/fix/PZ-8160-lint-errors?expand=1#diff-a8495babefd597f3b00988bdfdb640882faa10567eab564bda9489b387ec3c79):
Made detail optional since it's not used anywhere and not defined in the
constructor
-
[src/main/app/src/app/core/websocket/model/screen-event.ts](https://github.com/infonl/dimpact-zaakafhandelcomponent/compare/fix/PZ-8160-lint-errors?expand=1#diff-c6bcfd4dae33e7d2af97b465d13d17f2b79e85df02f42c65d848030f2e81a60c):
Made _timestamp optional since its not required in the constructor
params
-
[src/main/app/src/app/core/websocket/websocket.service.ts](https://github.com/infonl/dimpact-zaakafhandelcomponent/compare/fix/PZ-8160-lint-errors?expand=1#diff-f98ae1e14f4d58355e63f81fe09ac59bd74f3d64dd03e998471f2bfd6a7d6a9e):
Replaced the listeners and suspended property accordingly with
`Record<Types>`.

This PR further reduces the linting errors in the frontend of ZAC

Solves PZ-8160